Add simplified stateVector quantum simulator
Because the execution environment for the simulator is a little difficult to
build, we use a simplified version, instead. The goal is to reduce the overhead
of changing to the real simulator and of developing the PGO pass for the
simulator.

The simplified version contains the run_function which has a big switch cases
for different quantum gates. While the gates is not separated one by one but use
a large function, e.g., single_gate, for the configuration, we have to extract
the passed arguments in the runtime to identify which specific gate is called
and which qubits are involved in this gate.