Implement code injection by llvm pass
To profile the simulator, which has a large meta function called in the switch
cases, we leverage clang/llvm framework to implement a new pass as a plugin to
inject custom code (LLVM IR) into the target, i.e., the simulator.

The pass first traverse the basic blocks within functions, and insert code to
the call instructions. The arguments are collected and printed out to the stderr
currently.

Because the simulator has omp parallel pragma for the parallel computation, the
targets, e.g., single_gate, are called several times for processing a gate. To
avoid to collect repeated data, we add a if condition block for the inserted
code by checking current thread id is 0, which is just as the configuration
done. The timestamp is collected right before the called instructions by using
clock_gettime function with CLOCK_MONOTONIC argument. Finally, because of the
multithread environment, we use file lock to ensure the stderr output is not
interrupted by others before writing all info of a target.