Move to ModulePass
We originally use a FunctionPass for the pass, while the constructor is called
several times for each function. With the moving to ModulePass, the constructor
is only called when changing the module during the compilation which can reduce
the types initialization. Moreover, the register method is changed for the
ModulePass (EP_EarlyAsPossible causes program crash), and o0 cannot trigger the
pass.