Able to dump counters to the file
More than one counters will be increased if the gate is not single qubit gate,
so we change to use a vector to collect all qubit operands so that they can be
used as the index to load corresponding counter from the 2D array.

We increase the counters before the call instructions, and then it needs to be
dumped into a specific file for the counters before the program exit.