Fix correction of matrix from gate fusion output
The major problem is the ignoring qubit dependency of gates in reorder
phases. This will cause the incorrect output order of unitary gate,
leads to wrong execution result.
Add dependency check in the two reorder phases. "dependencyList" is used
to record a gate depend on other gates, constructed by
"constructDependencyList()". We record the previous gate operated on
some qubit by "gateOnQubit". When we iterate on "targetQubit" of each
gate, adding the previous gate on the qubit to the dependency list.
Using std::set as structure of dependency list due to avoiding
duplication.
Fix the wrong implementation of CX gate matrix. Due to the feature of CX
gate, we can simply apply XOR operation on matrix.