Fix incorrect RZZ implementation
Implement RZZ gate by finding the corresponding qubits. For example, if
we have a RZZ gate "RZZ 0 2 (theta)" and we want to expand to operate on 0
1 2 qubit.
Original RZZ on 2 qubits. Using "A" to replace the exponential things.
    __           __
00  | -A  0  0  0 |
01  |  0  A  0  0 |
10  |  0  0  A  0 |
11  |  0  0  0 -A |
    --           --

After expansion, we apply state by ignore the second qubit and take
first and third qubit merge back to two bit number. For example, state
101, we get 1 and 1, then merge to 0b11. Then use the numeber to
reference to the original RZZ matrix. We can get "-A" in this example.

 ignore
   |
   v     __                      __
0 |0| 0  | -A  0  0  0  0  0  0  0|
0 |0| 1  |  0  A  0  0  0  0  0  0|
0 |1| 0  |  0  0 -A  0  0  0  0  0|
0 |1| 1  |  0  0  0  A  0  0  0  0|
1 |0| 0  |  0  0  0  0  A  0  0  0|
1 |0| 1  |  0  0  0  0  0 -A  0  0|
1 |1| 0  |  0  0  0  0  0  0  A  0|
1 |1| 1  |  0  0  0  0  0  0  0 -A|
         --                      --